Restructure visiblity lambdas
