+++ visibility/email.go
@@ -1,4 +1,4 @@
-package main
+package visibility
 
 import (
 	"bytes"
